pyproject.toml
[tool.poetry]
name = "specctl"
version = "0.1.2"
description = ""
authors = ["Arvind Soni <arvindsoni@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
click = "^8.1.3"
PyYAML = "^6.0"
kubernetes = "^26.1.0"
boto3 = "^1.26.130"
pick = "^2.2.0"

[tool.poetry.scripts]
specctl = "specctl.specctl:transform"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
